Moved some utilities from glue.qt.qtutil to glue.utils.qt
