+++ glue/plugins/dendro_viewer/qt_widget.py
@@ -5,11 +5,12 @@ from glue import core
 from glue.plugins.dendro_viewer.client import DendroClient
 from glue.qt.glue_toolbar import GlueToolbar
 from glue.qt.mouse_mode import PickMode
-from glue.qt.qtutil import load_ui, nonpartial
+from glue.qt.qtutil import load_ui
 from glue.qt.widget_properties import (ButtonProperty, CurrentComboProperty,
                                        connect_bool_button, connect_current_combo)
 from glue.qt.widgets.data_viewer import DataViewer
 from glue.qt.widgets.mpl_widget import MplWidget, defer_draw
+from glue.utils import nonpartial
 
 
 class DendroWidget(DataViewer):
