+++ glue/qt/layer_artist_model.py
@@ -13,12 +13,12 @@ from __future__ import absolute_import, division, print_function
 
 from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui, QtCore
-from glue.qt.qtutil import (layer_artist_icon, nonpartial, PythonListModel)
-
-from glue.qt.mime import PyMimeData, LAYERS_MIME_TYPE
 from glue.clients.layer_artist import LayerArtistBase, LayerArtistContainer
-
+from glue.qt.qtutil import layer_artist_icon
+from glue.qt.mime import PyMimeData, LAYERS_MIME_TYPE
 from glue.qt.widgets.style_dialog import StyleDialog
+from glue.utils import nonpartial
+from glue.utils.qt import PythonListModel
 
 
 class LayerArtistModel(PythonListModel):
