+++ glue/qt/mouse_mode.py
@@ -25,6 +25,7 @@ from glue.core import roi
 from glue.qt import get_qapp
 from glue.qt import qt_roi
 from glue.qt.qtutil import get_icon, nonpartial, load_ui
+from glue.utils import nonpartial
 
 
 class MouseMode(object):
