+++ glue/qt/qt_roi.py
@@ -5,7 +5,7 @@ import numpy as np
 from glue.external.qt import QtCore
 from glue.external.qt import QtGui, QtCore
 from glue.core import roi
-from glue.qt.qtutil import mpl_to_qt4_color
+from glue.utils.qt import mpl_to_qt4_color
 
 
 class QtROI(object):
