+++ glue/qt/qtutil.py
@@ -7,63 +7,20 @@ from __future__ import absolute_import, division, print_function
 
 import os
 
-import numpy as np
 import pkg_resources
-from matplotlib import cm
-from matplotlib.colors import ColorConverter
 
 from glue.external.qt.QtCore import Qt
-from glue.external.qt import QtGui, QtCore, is_pyside
+from glue.external.qt import QtGui, QtCore
 from glue.external.axescache import AxesCache
 from glue import core
 from glue.qt import ui, icons
 from glue.qt.decorators import set_cursor
 from glue.qt.mime import PyMimeData, LAYERS_MIME_TYPE
-from glue.utils.qt import QMessageBoxPatched as QMessageBox
-
+from glue.utils.qt import (QMessageBoxPatched as QMessageBox, mpl_to_qt4_color,
+                           qt4_to_mpl_color, tint_pixmap, get_text)
 
 # We import nonpartial here for convenience
 from glue.utils import nonpartial
-
-
-def mpl_to_qt4_color(color, alpha=1.0):
-    """ Convert a matplotlib color stirng into a Qt QtGui.QColor object
-
-    :param color:
-       A color specification that matplotlib understands
-    :type color: str
-
-    :param alpha:
-       Optional opacity. Float in range [0,1]
-    :type alpha: float
-
-    * Returns *
-    A QtGui.QColor object representing color
-
-    :rtype: QtGui.QColor
-    """
-    if color in [None, 'none', 'None']:
-        return QtGui.QColor(0, 0, 0, 0)
-
-    cc = ColorConverter()
-    r, g, b = cc.to_rgb(color)
-    alpha = max(0, min(255, int(256 * alpha)))
-    return QtGui.QColor(r * 255, g * 255, b * 255, alpha)
-
-
-def qt4_to_mpl_color(color):
-    """
-    Convert a QtGui.QColor object into a string that matplotlib understands
-
-    Note: This ignores opacity
-
-    :param color: QtGui.QColor instance
-
-    *Returns*
-        A hex string describing that color
-    """
-    hexid = color.name()
-    return str(hexid)
 
 
 def data_wizard():
@@ -211,57 +168,6 @@ def edit_layer_label(layer):
         layer.label = str(label)
 
 
-def pick_item(items, labels, title="Pick an item", label="Pick an item",
-              default=0):
-    """ Prompt the user to choose an item
-
-    :param items: List of items to choose
-    :param labels: List of strings to label items
-    :param title: Optional widget title
-    :param label: Optional prompt
-
-    Returns the selected item, or None
-    """
-    choice, isok = QtGui.QInputDialog.getItem(None, title, label,
-                                              labels, current=default,
-                                              editable=False)
-    if isok:
-        index = labels.index(str(choice))
-        return items[index]
-
-
-def pick_class(classes, **kwargs):
-    """Prompt the user to pick from a list of classes using QT
-
-    :param classes: list of class objects
-    :param title: string of the prompt
-
-    Returns:
-       The class that was selected, or None
-    """
-    def _label(c):
-        try:
-            return c.LABEL
-        except AttributeError:
-            return c.__name__
-
-    choices = [_label(c) for c in classes]
-    return pick_item(classes, choices, **kwargs)
-
-
-def get_text(title='Enter a label'):
-    """Prompt the user to enter text using QT
-
-    :param title: Name of the prompt
-
-    *Returns*
-       The text the user typed, or None
-    """
-    result, isok = QtGui.QInputDialog.getText(None, title, title)
-    if isok:
-        return str(result)
-
-
 class GlueItemWidget(object):
 
     """ A mixin for QtGui.QListWidget/GlueTreeWidget subclasses, that
@@ -367,25 +273,6 @@ def layer_artist_icon(artist):
     return QtGui.QIcon(pm)
 
 
-def tint_pixmap(bm, color):
-    """Re-color a monochrome pixmap object using `color`
-
-    :param bm: QtGui.QBitmap instance
-    :param color: QtGui.QColor instance
-
-    :rtype: QtGui.QPixmap. The new pixma;
-    """
-    if bm.depth() != 1:
-        raise TypeError("Input pixmap must have a depth of 1: %i" % bm.depth())
-
-    image = bm.toImage()
-    image.setColor(1, color.rgba())
-    image.setColor(0, QtGui.QColor(0, 0, 0, 0).rgba())
-
-    result = QtGui.QPixmap.fromImage(image)
-    return result
-
-
 class GlueListWidget(GlueItemWidget, QtGui.QListWidget):
     pass
 
@@ -433,60 +320,6 @@ class GlueTabBar(QtGui.QTabBar):
         index = self.tabAt(event.pos())
         if index >= 0:
             self.rename_tab(index)
-
-
-def cmap2pixmap(cmap, steps=50):
-    """Convert a maplotlib colormap into a QtGui.QPixmap
-
-    :param cmap: The colormap to use
-    :type cmap: Matplotlib colormap instance (e.g. matplotlib.cm.gray)
-    :param steps: The number of color steps in the output. Default=50
-    :type steps: int
-
-    :rtype: QtGui.QPixmap
-    """
-    sm = cm.ScalarMappable(cmap=cmap)
-    sm.norm.vmin = 0.0
-    sm.norm.vmax = 1.0
-    inds = np.linspace(0, 1, steps)
-    rgbas = sm.to_rgba(inds)
-    rgbas = [QtGui.QColor(int(r * 255), int(g * 255),
-                          int(b * 255), int(a * 255)).rgba() for r, g, b, a in rgbas]
-    im = QtGui.QImage(steps, 1, QtGui.QImage.Format_Indexed8)
-    im.setColorTable(rgbas)
-    for i in range(steps):
-        im.setPixel(i, 0, i)
-    im = im.scaled(100, 100)
-    pm = QtGui.QPixmap.fromImage(im)
-    return pm
-
-
-def pretty_number(numbers):
-    """Convert a list of numbers into a nice list of strings
-
-    :param numbers: Numbers to convert
-    :type numbers: List or other iterable of numbers
-
-    :rtype: A list of strings
-    """
-    try:
-        return [pretty_number(n) for n in numbers]
-    except TypeError:
-        pass
-
-    n = numbers
-    if n == 0:
-        result = '0'
-    elif (abs(n) < 1e-3) or (abs(n) > 1e3):
-        result = "%0.3e" % n
-    elif abs(int(n) - n) < 1e-3 and int(n) != 0:
-        result = "%i" % n
-    else:
-        result = "%0.3f" % n
-    if result.find('.') != -1:
-        result = result.rstrip('0')
-
-    return result
 
 
 class RGBEdit(QtGui.QWidget):
@@ -860,235 +693,3 @@ if __name__ == "__main__":
 
     print(f.layer_visible)
     print(f.contrast_layer)
-
-
-class Worker(QtCore.QThread):
-    result = QtCore.Signal(object)
-    error = QtCore.Signal(object)
-
-    def __init__(self, func, *args, **kwargs):
-        """
-        Execute a function call on a different QtCore.QThread
-
-        :param func: The function object to call
-        :param args: arguments to pass to the function
-        :param kwargs: kwargs to pass to the function
-        """
-        super(Worker, self).__init__()
-        self.func = func
-        self.args = args
-        self.kwargs = kwargs
-
-    def run(self):
-        """
-        Invoke the function
-        Upon successful completion, the result signal will be fired
-        with the output of the function
-        If an exception occurs, the error signal will be fired with
-        the result form sys.exc_infno()
-        """
-        try:
-            result = self.func(*self.args, **self.kwargs)
-            self.result.emit(result)
-        except:
-            import sys
-            self.error.emit(sys.exc_info())
-
-
-def update_combobox(combo, labeldata):
-    """
-    Redefine the items in a combobox
-
-    Parameters
-    ----------
-    widget : QComboBox
-       The widget to update
-    labeldata : sequence if N (label, data) tuples
-       The combobox will contain N items with the appropriate
-       labels, and data set as the userData
-
-    Returns
-    -------
-    combo : QComboBox
-        The updated input
-
-    Notes
-    -----
-    If the current userData in the combo box matches
-    any of labeldata, that selection will be retained.
-    Otherwise, the first item will be selected.
-
-    QtCore.Signals are disabled while the combo box is updated
-
-    combo is modified inplace
-    """
-    combo.blockSignals(True)
-    idx = combo.currentIndex()
-    if idx >= 0:
-        current = combo.itemData(idx)
-    else:
-        current = None
-
-    combo.clear()
-    index = 0
-    for i, (label, data) in enumerate(labeldata):
-        combo.addItem(label, userData=data)
-        if data is current:
-            index = i
-    combo.blockSignals(False)
-    combo.setCurrentIndex(index)
-
-    # We need to force emit this, otherwise if the index happens to be the
-    # same as before, even if the data is different, callbacks won't be
-    # called.
-    if idx == index or idx == -1:
-        combo.currentIndexChanged.emit(index)
-
-    return combo
-
-
-class PythonListModel(QtCore.QAbstractListModel):
-
-    """
-    A Qt Model that wraps a python list, and exposes a list-like interface
-
-    This can be connected directly to multiple QListViews, which will
-    stay in sync with the state of the container.
-    """
-
-    def __init__(self, items, parent=None):
-        """
-        Create a new model
-
-        Parameters
-        ----------
-        items : list
-            The initial list to wrap
-        parent : QObject
-            The model parent
-        """
-        super(PythonListModel, self).__init__(parent)
-        self.items = items
-
-    def rowCount(self, parent=None):
-        """Number of rows"""
-        return len(self.items)
-
-    def headerData(self, section, orientation, role):
-        """Column labels"""
-        if role != Qt.DisplayRole:
-            return None
-        return "%i" % section
-
-    def row_label(self, row):
-        """ The textual label for the row"""
-        return str(self.items[row])
-
-    def data(self, index, role):
-        """Retrieve data at each index"""
-        if not index.isValid():
-            return None
-        if role == Qt.DisplayRole or role == Qt.EditRole:
-            return self.row_label(index.row())
-        if role == Qt.UserRole:
-            return self.items[index.row()]
-
-    def setData(self, index, value, role):
-        """
-        Update the data in-place
-
-        Parameters
-        ----------
-        index : QtCore.QModelIndex
-            The location of the change
-        value : object
-            The new value
-        role : QEditRole
-            Which aspect of the model to update
-        """
-        if not index.isValid():
-            return False
-
-        if role == Qt.UserRole:
-            row = index.row()
-            self.items[row] = value
-            self.dataChanged.emit(index, index)
-            return True
-
-        return super(PythonListModel, self).setDdata(index, value, role)
-
-    def removeRow(self, row, parent=None):
-        """
-        Remove a row from the table
-
-        Parameters
-        ----------
-        row : int
-            Row to remove
-
-        Returns
-        -------
-        successful : bool
-        """
-        if row < 0 or row >= len(self.items):
-            return False
-
-        self.beginRemoveRows(QtCore.QModelIndex(), row, row)
-        self._remove_row(row)
-        self.endRemoveRows()
-        return True
-
-    def pop(self, row=None):
-        """
-        Remove and return an item (default last item)
-
-        Parameters
-        ----------
-        row : int (optional)
-            Which row to remove. Default=last
-
-        Returns
-        --------
-        popped : object
-        """
-        if row is None:
-            row = len(self) - 1
-        result = self[row]
-        self.removeRow(row)
-        return result
-
-    def _remove_row(self, row):
-        # actually remove data. Subclasses can override this as needed
-        self.items.pop(row)
-
-    def __getitem__(self, row):
-        return self.items[row]
-
-    def __setitem__(self, row, value):
-        index = self.index(row)
-        self.setData(index, value, role=Qt.UserRole)
-
-    def __len__(self):
-        return len(self.items)
-
-    def insert(self, row, value):
-        self.beginInsertRows(QtCore.QModelIndex(), row, row)
-        self.items.insert(row, value)
-        self.endInsertRows()
-        self.rowsInserted.emit(self.index(row), row, row)
-
-    def append(self, value):
-        row = len(self)
-        self.insert(row, value)
-
-    def extend(self, values):
-        for v in values:
-            self.append(v)
-
-    def set_list(self, values):
-        """
-        Set the model to a new list
-        """
-        self.beginResetModel()
-        self.items = values
-        self.endResetModel()
