+++ glue/qt/simpleforms.py
@@ -3,8 +3,7 @@ from __future__ import absolute_import, division, print_function
 from glue.external.qt.QtCore import QObject, Signal
 from glue.external.qt import QtGui
 from glue.core.simpleforms import IntOption, FloatOption, BoolOption
-from glue.qt.qtutil import nonpartial
-
+from glue.utils import nonpartial
 
 _dispatch = {}
