+++ glue/qt/widget_properties.py
@@ -24,7 +24,7 @@ from functools import partial
 from glue.external.six.moves import reduce
 from glue.external.qt import QtGui
 from glue.core.callback_property import add_callback
-from glue.qt.qtutil import pretty_number
+from glue.utils.array import pretty_number
 
 
 class WidgetProperty(object):
