+++ glue/qt/widgets/data_slice_widget.py
@@ -4,10 +4,11 @@ from functools import partial
 from collections import Counter
 
 from glue.external.qt import QtGui, QtCore
-from glue.qt.qtutil import nonpartial, load_ui
+from glue.qt.qtutil import load_ui
 from glue.qt.widget_properties import (TextProperty,
                                        ValueProperty,
                                        CurrentComboProperty)
+from glue.utils import nonpartial
 
 
 class SliceWidget(QtGui.QWidget):
