+++ glue/qt/widgets/style_dialog.py
@@ -2,8 +2,8 @@ from __future__ import absolute_import, division, print_function
 
 from glue.external.qt.QtCore import Qt
 from glue.external.qt import QtGui, QtCore
-from glue.qt.qtutil import (mpl_to_qt4_color, symbol_icon, POINT_ICONS,
-                            qt4_to_mpl_color)
+from glue.qt.qtutil import POINT_ICONS, symbol_icon
+from glue.utils.qt import mpl_to_qt4_color, qt4_to_mpl_color
 
 
 class ColorWidget(QtGui.QLabel):
