+++ glue/qt/widgets/subset_facet.py
@@ -5,9 +5,11 @@ from matplotlib import cm
 
 from glue.external.qt import QtGui
 from glue.core.util import colorize_subsets, facet_subsets, Pointer
-from glue.qt.qtutil import pretty_number, cmap2pixmap, load_ui
+from glue.qt.qtutil import load_ui
 from glue.qt.widget_properties import (ButtonProperty, FloatLineProperty,
                                        ValueProperty)
+from glue.utils.array import pretty_number
+from glue.utils.qt import cmap2pixmap
 
 
 class SubsetFacet(object):
