+++ glue/utils/qt/__init__.py
@@ -2,3 +2,8 @@ from __future__ import absolute_import, division, print_function
 
 from .autocomplete_widget import *
 from .qmessagebox_widget import *
+from .colors import *
+from .dialogs import *
+from .helpers import *
+from .python_list_model import *
+from .threading import *
